requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
colorama==0.4.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
isort==5.5.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.5
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pypinyin==0.39.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
